requirements.txt
pandas==2.3.3
numpy==2.3.5
SQLAlchemy==2.0.44
psycopg2-binary==2.9.11
scikit-learn==1.7.2
plotly==6.5.0
dash==3.3.0
dash-bootstrap-components==2.0.4
pytest==9.0.1
python-dotenv>=1.0.0,<2.0.0
